+++ modules/AWSPowerShell/Cmdlets/Lambda/AWS.Tools.Lambda.Completers.psm1
@@ -178,6 +178,13 @@ $LM_Completers = {
             break
         }
 
+        # Amazon.Lambda.UpdateRuntimeOn
+        "Write-LMRuntimeManagementConfig/UpdateRuntimeOn"
+        {
+            $v = "Auto","FunctionUpdate","Manual"
+            break
+        }
+
 
     }
 
@@ -200,6 +207,7 @@ $LM_map = @{
     "SnapStart_ApplyOn"=@("Publish-LMFunction","Update-LMFunctionConfiguration")
     "StartingPosition"=@("New-LMEventSourceMapping")
     "TracingConfig_Mode"=@("Publish-LMFunction","Update-LMFunctionConfiguration")
+    "UpdateRuntimeOn"=@("Write-LMRuntimeManagementConfig")
 }
 
 _awsArgumentCompleterRegistration $LM_Completers $LM_map
@@ -284,6 +292,7 @@ $LM_SelectMap = @{
                "Get-LMLayerVersionPolicy",
                "Get-LMPolicy",
                "Get-LMProvisionedConcurrencyConfig",
+               "Get-LMRuntimeManagementConfig",
                "Invoke-LMFunction",
                "Invoke-LMFunctionAsync",
                "Get-LMAliasList",
@@ -304,6 +313,7 @@ $LM_SelectMap = @{
                "Write-LMFunctionConcurrency",
                "Write-LMFunctionEventInvokeConfig",
                "Write-LMProvisionedConcurrencyConfig",
+               "Write-LMRuntimeManagementConfig",
                "Remove-LMLayerVersionPermission",
                "Remove-LMPermission",
                "Add-LMResourceTag",