+++ modules/AWSPowerShell/Cmdlets/Lambda/Basic/New-LMAlias-Cmdlet.cs
@@ -28,7 +28,7 @@ using Amazon.Lambda.Model;
 namespace Amazon.PowerShell.Cmdlets.LM
 {
     /// <summary>
-    /// Creates an <a href="https://docs.aws.amazon.com/lambda/latest/dg/versioning-aliases.html">alias</a>
+    /// Creates an <a href="https://docs.aws.amazon.com/lambda/latest/dg/configuration-aliases.html">alias</a>
     /// for a Lambda function version. Use aliases to provide clients with a function identifier
     /// that you can update to invoke a different version.
     ///