+++ CHANGELOG.md
@@ -1,4 +1,10 @@
-﻿### 4.1.254 (2023-01-20 22:04Z)
+﻿### 4.1.255 (2023-01-23 22:07Z)
+  * AWS Tools for PowerShell now use AWS .NET SDK 3.7.466.0 and leverage its new features and improvements. Please find a description of the changes at https://github.com/aws/aws-sdk-net/blob/master/SDK.CHANGELOG.md.
+  * Amazon Lambda
+    * Added cmdlet Get-LMRuntimeManagementConfig leveraging the GetRuntimeManagementConfig service API.
+    * Added cmdlet Write-LMRuntimeManagementConfig leveraging the PutRuntimeManagementConfig service API.
+
+### 4.1.254 (2023-01-20 22:04Z)
   * AWS Tools for PowerShell now use AWS .NET SDK 3.7.465.0 and leverage its new features and improvements. Please find a description of the changes at https://github.com/aws/aws-sdk-net/blob/master/SDK.CHANGELOG.md.
 
 ### 4.1.253 (2023-01-20 02:05Z)